+++ src/objects.js
@@ -94,7 +94,7 @@ embedMetadataPNG, SnapExtensions*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2022-November-15';
+modules.objects = '2022-November-16';
 
 var SpriteMorph;
 var StageMorph;
@@ -4000,6 +4000,12 @@ SpriteMorph.prototype.unflashScope = function (varName, isGlobal) {
 };
 
 SpriteMorph.prototype.visibleScopeFor = function (varName, isGlobal) {
+    // private - answer an array of all my syntax elements within the lexical
+    // scope of a given variable name, so they can be highlighted in the IDE.
+    // Note: This is optimized to only answer the *visible* blocks, if you
+    // want to get the full collection of affected elements use
+    // Block >> fullScopeFor(varName) instead.
+
     var elements = [];
 
     if (isGlobal && contains(this.variables.names(), varName)) {